requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0

# Auth dependencies - Kullanıcı giriş/çıkış sistemi için
sqlalchemy==2.0.23  # Veritabanı ORM
python-jose[cryptography]==3.3.0  # JWT token işlemleri
bcrypt>=4.0.1  # Bcrypt algoritması (chromadb ile uyumlu)
pydantic[email]

# CHAIN SYSTEM - LangChain dependencies (uyumlu versiyonlar)
langchain==0.3.7
langchain-openai==0.2.8
langchain-google-genai==2.0.4
openai>=1.54.0,<2.0.0
google-generativeai>=0.8.0

# RAG dependencies
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
PyPDF2

# Summarizer (T5-small via transformers)
transformers>=4.40.0,<5.0.0  # Uyumlu versiyon aralığı
sentencepiece>=0.2.0
safetensors>=0.4.5

# LoRA Fine-Tuning dependencies
torch --index-url https://download.pytorch.org/whl/cu121
peft>=0.6.0,<1.0.0  # Parameter Efficient Fine-Tuning (LoRA için)
datasets>=2.14.0  # Hugging Face datasets
accelerate>=0.24.0  # Training hızlandırma ve GPU yönetimi (CPU'da da çalışır)
numpy>=1.24.0,<2.0.0  # Numerik işlemler (genellikle torch ile gelir ama ek güvenlik için)
huggingface_hub[cli]>=0.20.0

# NOT: GPU kullanmak için PyTorch CUDA versiyonu kurulmalıdır:
# CUDA 12.1: pip install torch --index-url https://download.pytorch.org/whl/cu121
# CUDA 11.8: pip install torch --index-url https://download.pytorch.org/whl/cu118
# Script otomatik olarak GPU varsa kullanır, yoksa CPU'da çalışır (daha yavaş)